+++ src/apis/followData.js
@@ -1,9 +1,9 @@
-import { apiHelper } from './../utils/helpers'
+import { apiHelper } from '../utils/helpers'
 const getToken = () => localStorage.getItem('token')
 
 export default {
-  getUserProfile(userId) {
-    return apiHelper.get(`/users/${userId}`, { headers: { Authorization: `Bearer ${getToken()}` } })
+  getPopularList() {
+    return apiHelper.get(`/users/top`, { headers: { Authorization: `Bearer ${getToken()}` } })
   },
   getUserFollowings(userId) {
     return apiHelper.get(`/users/${userId}/followings`, { headers: { Authorization: `Bearer ${getToken()}` } })
