+++ src/apis/tweetData.js
@@ -5,6 +5,9 @@ export default {
   getTweetCards() {
     return apiHelper.get(`/tweets`, { headers: { Authorization: `Bearer ${getToken()}` } })
   },
+  postTweet({ description }) {
+    return apiHelper.post(`/tweets`, { description }, { headers: { Authorization: `Bearer ${getToken()}` } })
+  },
   addTweetLike(id) {
     return apiHelper.post(`/tweets/${id}/like`, null, { headers: { Authorization: `Bearer ${getToken()}` } })
   },
