+++ src/components/FollowCards.vue
@@ -37,7 +37,7 @@
 
 <script>
 import { emptyImageFilter } from "../utils/mixins";
-import userFollowAPI from "./../apis/userFollow";
+import userFollowAPI from "./../apis/followData";
 import { Toast } from "./../utils/helpers";
 
 export default {
