+++ src/components/ReplyListTweet.vue
@@ -71,7 +71,7 @@
 <script>
 import { formatDateFilter } from "./../utils/mixins";
 import { emptyImageFilter } from "../utils/mixins";
-import tweetAPI from "../apis/tweetCards";
+import tweetAPI from "../apis/tweetData";
 import { Toast } from "./../utils/helpers";
 import { mapState } from "vuex";
